+++ ESFramework/UIKit/UIApplication+ESHelper.m
@@ -7,12 +7,6 @@
 //
 
 #import "UIApplication+ESAdditions.h"
-#import <objc/runtime.h>
-#import "ESMacros.h"
-
-NSString *const ESMultitaskingBackgroundTaskName = @"ESMultitaskingBackgroundTask";
-
-ESDefineAssociatedObjectKey(multitaskingBackgroundTaskIdentifier)
 
 @implementation UIApplication (ESHelper)
 
@@ -28,44 +22,6 @@ ESDefineAssociatedObjectKey(multitaskingBackgroundTaskIdentifier)
     } else {
         printf("UIApplication no longer responds \"_performMemoryWarning\" selector.\n");
         exit(1);
-    }
-}
-
-- (UIBackgroundTaskIdentifier)multitaskingBackgroundTaskIdentifier
-{
-    NSNumber *value = objc_getAssociatedObject(self, multitaskingBackgroundTaskIdentifierKey);
-    return value ? [value unsignedIntegerValue] : UIBackgroundTaskInvalid;
-}
-
-- (void)setMultitaskingBackgroundTaskIdentifier:(UIBackgroundTaskIdentifier)identifier
-{
-    objc_setAssociatedObject(self, multitaskingBackgroundTaskIdentifierKey,
-                             @(identifier), OBJC_ASSOCIATION_RETAIN_NONATOMIC);
-}
-
-- (BOOL)isMultitaskingEnabled
-{
-    return self.multitaskingBackgroundTaskIdentifier != UIBackgroundTaskInvalid;
-}
-
-- (UIBackgroundTaskIdentifier)enableMultitasking
-{
-    if (!self.isMultitaskingEnabled) {
-        self.multitaskingBackgroundTaskIdentifier =
-            [self beginBackgroundTaskWithName:ESMultitaskingBackgroundTaskName expirationHandler:^{
-                [self disableMultitasking];
-                [self enableMultitasking];
-            }];
-    }
-
-    return self.multitaskingBackgroundTaskIdentifier;
-}
-
-- (void)disableMultitasking
-{
-    if (self.isMultitaskingEnabled) {
-        [self endBackgroundTask:self.multitaskingBackgroundTaskIdentifier];
-        self.multitaskingBackgroundTaskIdentifier = UIBackgroundTaskInvalid;
     }
 }